Changed the Exception stringstream to be a defined stream rather than a pointer to a stream to avoid a nasty segmentation fault that could occur in a rather bizarrre use case in which there was a particular type of error that occured after a Problem object had been created and destroyed within another Problem.
Updated Telements to report information about the number of nodes on their
faces

Modified the unstructured two-layer soluble surfactant problem so that it
now works under refinement over the periodic boundary.
This is still not fully tested, but compiles and runs for a while,
so I want to check it in.